+++ pkg/utils/browser.go
@@ -8,6 +8,6 @@ import (
 
 func TryOpenBrowser(url string) {
 	if err := open.Run(url); err != nil {
-		fmt.Printf("Browse to: %s", url)
+		fmt.Printf("Browse to: %s\n", url)
 	}
 }